Improve sample
