+++ vitamio-sample/res/layout/videoview.xml
@@ -4,17 +4,23 @@
     android:layout_height="match_parent"
     android:orientation="vertical" >
 
-    <io.vov.vitamio.widget.CenterLayout
-        android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        android:orientation="vertical" >
+    <io.vov.vitamio.widget.VideoView
+        android:id="@+id/surface_view"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content" />
 
-        <io.vov.vitamio.widget.VideoView
-            android:id="@+id/surface_view"
-            android:layout_width="match_parent"
-            android:layout_height="match_parent"
-            android:layout_centerHorizontal="true"
-            android:layout_centerVertical="true" />
-    </io.vov.vitamio.widget.CenterLayout>
+    <Button
+        android:id="@+id/button1"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:text="@string/openvideo1" 
+        android:onClick="openVideo"/>
+
+    <Button
+        android:id="@+id/button2"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:text="@string/openvideo2"
+        android:onClick="openVideo2" />
 
 </LinearLayout>
